[split]  util-eval: fix for when class files are on the classpath directly
fix for when scala.tool.nsc.Interpreter and scala.ScalaObject class
files are on the classpath directly instead of in jar. fix for when
multiple class loaders are used. i needed both these fixes to use
Eval with 'hadoop jar' and a fat jar since hadoop unpacks the jar
to a temporary directory and puts the class files directly on the
classpath by introducing a new class loader

Signed-off-by: Rao Fu <rao@twitter.com>

RB_ID=142194