allow DirectoryDependencies on BuildBuilds
this fixes a bug where finalBuild would over eagerly go down all the way
down to the outermost build instead of stopping at the one requested.
Now it checks the new argument and stops there.

This is necessary to allow having one build depend on another build in
order to embed it in a type-safe way and have access to it’s tasks.