ISSUE-558 Correct ordering of imports
In order to resolve imports, one need to :
- resolve the properties of the xml declaration of the import. This
means that the parents have to be resolved first.
- resolve the properties of the dependencies contained by the import.
This means that when looking up values in the xml, you potentially have
to resolve the transitive poms of the import first.

This therefore parameterise the "lookup" and "properties" method in
order to take the scope into consideration. Typically the argument will
be "transitivePom" or "transitivePom and transitiveImports"

Also changes the order of how the transitivePom and imports are mixed
together.